python_version/requirements.txt
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=9.0.0
